[REFACTOR/#3] 멀티뷰 아이템 이름 변경에 따른 리팩토링 코드 추가
