+++ app/src/main/java/org/android/go/sopt/ui/main/home/adapter/MultiViewHolder.kt
@@ -2,24 +2,25 @@ package org.android.go.sopt.ui.main.home.adapter
 
 import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.RecyclerView
-import org.android.go.sopt.databinding.ItemImageBinding
-import org.android.go.sopt.databinding.ItemTextBinding
+import org.android.go.sopt.databinding.ItemHeaderBinding
+import org.android.go.sopt.databinding.ItemRepoBinding
+import org.android.go.sopt.ui.main.home.adapter.model.MultiViewItem
 
 /** sealed class는 추상 클래스로 자식 클래스의 종류를 제한할 수 있다. */
 sealed class MultiViewHolder<E : MultiViewItem>(
     binding: ViewDataBinding
 ) : RecyclerView.ViewHolder(binding.root) {
 
-    class TextViewHolder(
-        private val binding: ItemTextBinding
+    class HeaderViewHolder(
+        private val binding: ItemHeaderBinding
     ) : MultiViewHolder<MultiViewItem.Header>(binding) {
         override fun bind(item: MultiViewItem.Header) {
             binding.title = item.title
         }
     }
 
-    class ImageViewHolder(
-        private val binding: ItemImageBinding
+    class RepoViewHolder(
+        private val binding: ItemRepoBinding
     ) : MultiViewHolder<MultiViewItem.Repo>(binding) {
         override fun bind(item: MultiViewItem.Repo) {
             with(binding){
