+++ app/src/main/java/org/android/go/sopt/ui/main/home/adapter/MultiViewType.kt
@@ -1,6 +1,6 @@
 package org.android.go.sopt.ui.main.home.adapter
 
 enum class MultiViewType {
-    TEXT,
-    IMAGE
+    HEADER_TYPE,
+    REPO_TYPE
 }
